feat: added Swal alert

+++ src/components/lists/ListContainer.jsx
@@ -3,6 +3,7 @@ import style from './ListContainer.module.scss';
 import axios from 'axios';
 import url from '../../BackendURL';
 import List from './List';
+import Swal from 'sweetalert2';
 
 function ListContainer() {
 
@@ -25,11 +26,26 @@ function ListContainer() {
    * deletes a list by its id
    */
   const deleteList = async (list) => {
-    await axios.delete(url + '/api/lists/' + list.id)
-      .then((res) => {
-        console.log(res.data);
-        getLists();
-      });
+    Swal.fire({
+      position: 'center',
+      icon: 'warning',
+      iconColor: '#ff6a6a',
+      title: 'Bist du sicher, dass die gesamte Liste inkl. aller Todos gelöscht werden soll',
+      showConfirmButton: true,
+      confirmButtonText: 'Ja',
+      confirmButtonColor: '#bef983',
+      showCancelButton: true,
+      cancelButtonText: 'Nein',
+      cancelButtonColor: '#ff6a6a',
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        await axios.delete(url + '/api/lists/' + list.id)
+          .then((res) => {
+            console.log(res.data);
+            getLists();
+          });
+      }
+    });
   }
 
   return (
